Re-org beforeValue and testCase helpers
