+++ packages/beacon-state-transition/test/perf/altair/epoch/epoch.test.ts
@@ -7,8 +7,7 @@ import {
   CachedBeaconStateAltair,
   beforeProcessEpoch,
 } from "../../../../src";
-import {getNetworkCachedState} from "../../../utils/testFileCache";
-import {beforeValue, LazyValue} from "../../util";
+import {getNetworkCachedState, beforeValue, LazyValue} from "../../../utils";
 import {StateEpoch} from "../../types";
 import {altairState} from "../../params";
